Move governance checks out of testing package
